fix trailing zeroes and pangolin section
